requirements.txt
fastapi==0.70
google-cloud-storage==1.42.3
pydantic
uvicorn